requirements.txt
pandas
sentence-transformers
faiss-cpu
openai
fuzzywuzzy
python-Levenshtein
